Benchmarks (#46)
* Fix insertion of colored images

* Added face recognition positive and negative viewer

* Added functionality to do benchmarking with sklearn images

+++ benchmarks/FaceRecognitionGUI/benchmark.py
@@ -10,6 +10,7 @@
 #System Libraries
 import sys
 import os
+from pathlib import Path
 
 # debugging
 from inspect import currentframe, getframeinfo
@@ -101,8 +102,10 @@ class benchRecog():
         self.pW = None          # progress window
 
         self.userLimit = 5
-        self.trainPictureAmount = 20
-        self.recogPictureAmount = 10
+        # TODO: Let the user set it in the future
+        # set the picture amount
+        self.trainPictureAmount = 3
+        self.recogPictureAmount = 3
 
         # initialize timers for respective tests
         self.TPUserTimer = BVU.UserTimer()
@@ -113,6 +116,8 @@ class benchRecog():
         self.OFTNUserTimer = BVU.UserTimer()
         self.MixedUserTimer = BVU.UserTimer()
         self.FaceRecog2023UserTimer = BVU.UserTimer()
+        self.FaceRecog2023PositiveUserTimer = BVU.UserTimer()
+        self.FaceRecog2023NegativeUserTimer = BVU.UserTimer()
     
         self.exitFlag = False
         self.cv2Inst = cv2Recog()
@@ -158,13 +163,32 @@ class benchRecog():
                 self.exitFlag = True
                 return
 
-        resizedImgs = []
-        self.image_shape = imgs[0].shape
+        # Set shape of the images
+        # This can also be determinted in another way
+        # self.image_shape = self.master.imShape
+        self.image_shape = (125, 94)
 
         # Resize images
         # Since WiRe Algo needs all images to be the same height and width
-        for img in imgs:
-            resizedImgs.append(img.reshape((self.image_shape[1] * self.image_shape[0])))
+        resizedImgs = []
+        for img_index, img in enumerate(imgs):
+            if len(img.shape) == 3 and img.shape[2] == 3: 
+                img = cv2.cvtColor(img.astype("uint8"), cv2.COLOR_BGR2GRAY)
+
+            to_convert = cv2.normalize(
+                src=img, 
+                dst=None, 
+                alpha=0, 
+                beta=255, 
+                norm_type=cv2.NORM_MINMAX, 
+                dtype=cv2.CV_8U
+            )
+            resizedImg = cv2.resize(
+                to_convert.astype("uint8"),
+                dsize=(self.image_shape[1], self.image_shape[0]), 
+                interpolation=cv2.INTER_CUBIC
+            ).flatten()
+            resizedImgs.append(resizedImg)
 
         # update progress window
         if self.pW:
@@ -197,6 +221,7 @@ class benchRecog():
             data = self.df.loc[index]
             username = list(data['name'])[0]
             user_ = userRecog(username)
+            user_.imgShape = self.image_shape
 
             dataRecog = np.asarray(data['img'].iloc[:self.recogPictureAmount])
             dataTrain = np.asarray(data['img'].iloc[self.recogPictureAmount:self.trainPictureAmount+self.recogPictureAmount])
@@ -218,10 +243,11 @@ class benchRecog():
             self.pW.finProgress("benchProg")
 
         # reset users
-        self.users = self.master.DbUsers # TODO: This should be fixed (Don't know why they did that)
         self.dbUsers = self.master.DbUsers
 
-        self.image_shape = self.master.imShape
+        # TODO: Give the user an option to do this
+        # Comment it in if you want to use the database for the benchmark tests
+        # self.users = self.master.DbUsers
 
     def modified_project_faces(self, pcs: np.ndarray, images: np.ndarray, mean_data: np.ndarray) -> np.ndarray:
         """
@@ -976,42 +1002,46 @@ recogScore values:\n {recogScores}
 
         return recogScores
 
-    def face_recog_2023_run_benchmark(self):
-        def execute_algorithm(train_imgs , test_imgs):
-            scores = np.empty((len(train_imgs), len(test_imgs)))
-            scores[:] = np.nan
-            faceRecognitionInstance = FaceReco_class.FaceReco()
+    def _execute_face_recog_2023_algo(self, train_imgs , test_imgs):
+        if (len(train_imgs) == 0) or (len(test_imgs) == 0):
+            return np.array([])
+        scores = np.empty((len(train_imgs), len(test_imgs)))
+        scores[:] = np.nan
+        faceRecognitionInstance = FaceReco_class.FaceReco()
 
-            for train_index, train_img in enumerate(train_imgs):
-                train_img = train_img.reshape((self.image_shape[0],self.image_shape[1]))
-                train_img = train_img.astype("uint8")
+        for train_index, train_img in enumerate(train_imgs):
+            train_img = train_img.reshape((self.image_shape[0],self.image_shape[1]))
+            train_img = train_img.astype("uint8")
 
-                # calculate encoding
-                pil_img = Image.fromarray(train_img)
-                pil_img.save("train.jpg")
-                img_train = face_recognition.load_image_file("train.jpg")
-                image_encoding = face_recognition.face_encodings(img_train)
+            # calculate encoding
+            pil_img = Image.fromarray(train_img)
+            pil_img.save("train.jpg")
+            img_train = face_recognition.load_image_file("train.jpg")
+            image_encoding = face_recognition.face_encodings(img_train)
 
-                for test_index, test_img in enumerate(test_imgs):
-                    ## it is possible that the encoding can't be calculated
-                    if image_encoding == []:
-                        # NOTE: The score is set to 1, because this is the maximum distance
-                        # and the picture is assumed to be completely different
-                        score = [1]
-                    else:
-                        # calculate distance
-                        test_img = test_img.reshape((self.image_shape[0],self.image_shape[1]))
-                        test_img = test_img.astype("uint8")
-                        pil_img = Image.fromarray(test_img)
-                        pil_img.save("test.jpg")
-                        img_test = face_recognition.load_image_file("test.jpg")
+            for test_index, test_img in enumerate(test_imgs):
+                ## it is possible that the encoding can't be calculated
+                if image_encoding == []:
+                    # NOTE: The score is set to 1, because this is the maximum distance
+                    # and the picture is assumed to be completely different
+                    score = [1]
+                else:
+                    # calculate distance
+                    test_img = test_img.reshape((self.image_shape[0],self.image_shape[1]))
+                    test_img = test_img.astype("uint8")
+                    pil_img = Image.fromarray(test_img)
+                    pil_img.save("test.jpg")
+                    img_test = face_recognition.load_image_file("test.jpg")
 
-                        _, score = faceRecognitionInstance.photo_to_photo(image_encoding[0], img_test)
-                    scores[train_index][test_index] = score[0]
+                    _, score = faceRecognitionInstance.photo_to_photo(image_encoding[0], img_test)
+                scores[train_index][test_index] = score[0]
+        if Path("test.jpg").is_file():
             os.remove("test.jpg")
+        if Path("train.jpg").is_file():
             os.remove("train.jpg")
-            return scores
+        return scores
 
+    def face_recog_2023_run_benchmark(self):
         LABEL_FOR_PROGRESS_OF_TEST = "FaceRecog2023Prog"
 
         if self.pW:
@@ -1041,7 +1071,7 @@ recogScore values:\n {recogScores}
             if self.root:
                 self.root.update()
 
-            if len(user_.trainPictures) == 0:
+            if len(user_.trainPictures) == 0 or len(user_.recogPictures) == 0:
                 print("WARNING: Training pictures empty!")
                 continue
 
@@ -1057,7 +1087,7 @@ recogScore values:\n {recogScores}
             imgs_test = user_.recogPictures.copy()
 
             ## calculate scores
-            scores = execute_algorithm(imgs_train, imgs_test)
+            scores = self._execute_face_recog_2023_algo(imgs_train, imgs_test)
 
             ## Prepare Score Dataframe
             user_.recogScores = None
@@ -1083,7 +1113,7 @@ recogScore values:\n {recogScores}
             imgs_test = user_.recogPictures.copy()
 
             ## calculate scores
-            scores = execute_algorithm(imgs_train, imgs_test)
+            scores = self._execute_face_recog_2023_algo(imgs_train, imgs_test)
 
             ## Prepare Score Dataframe
             user_.recogScores = None
@@ -1134,6 +1164,171 @@ recogScore values:\n {recogScores}
         print(
 f"""
 Description: This is the output of the true negative test that inserts data.
+File: {frame_info.filename}
+Line: {frame_info.lineno}
+recogScore values:\n {recogScores}
+"""
+        )
+
+        return recogScores
+
+    def face_recog_2023_run_positive(self):
+        LABEL_FOR_PROGRESS_OF_TEST = "FaceRecog2023PositiveProg"
+
+        if self.pW:
+            self.pW.createProgressbar(LABEL_FOR_PROGRESS_OF_TEST)
+
+        userCount = len(self.users)
+        userFin = 0
+
+        testImageIndices = []
+        recogImageIndices = []
+        recogImageScores = []
+        usernames = []
+        recogUsernames = []
+
+        # reset previously used resources
+        self.FaceRecog2023PositiveUserTimer.clear()
+
+        for user_index,user_ in enumerate(self.users):
+            if self.pW:
+                self.pW.update(
+                    LABEL_FOR_PROGRESS_OF_TEST,
+                    "[FaceRecog2023Positive] Benchmarking User: {}/{}...".format(user_.username,self.userLimit),
+                    userFin/userCount * 100
+                )
+
+            if self.root:
+                self.root.update()
+
+            if len(user_.trainPictures) == 0 or len(user_.recogPictures) == 0:
+                print("WARNING: Training pictures empty!")
+                continue
+
+            self.FaceRecog2023PositiveUserTimer.startTimer(user_)
+
+            ## Fetch User pictures
+            imgs_train = user_.trainPictures.copy() # Images are flattened
+            imgs_test = user_.recogPictures.copy()
+
+            ## calculate scores
+            scores = self._execute_face_recog_2023_algo(imgs_train, imgs_test)
+
+            ## Prepare Score Dataframe
+            user_.recogScores = None # TODO: this is not used somehow
+            for testImageIndex in range(scores.shape[1]):
+                recognisedImageIndex = np.argmin(scores[:, testImageIndex])
+                recognisedImageScore = np.min(scores[:, testImageIndex])
+                testImageIndices.append(testImageIndex)
+                recogImageIndices.append(recognisedImageIndex)
+                recogImageScores.append(recognisedImageScore)
+                usernames.append(user_.username)
+                recogUsernames.append(user_.username)
+
+            userFin += 1
+            self.FaceRecog2023PositiveUserTimer.endTimer(user_)
+
+        if self.pW:
+            self.pW.finProgress(LABEL_FOR_PROGRESS_OF_TEST)
+
+        recogScores = pd.DataFrame(
+            data = {
+                'testImageIndex':  testImageIndices,
+                'username':        usernames, 
+                'recogImageIndex': recogImageIndices,
+                'recogUsername' :  recogUsernames,
+                'recogScore' :     recogImageScores
+            }).set_index('recogImageIndex').sort_values(by='testImageIndex')
+
+        frame_info = getframeinfo(currentframe())
+        print(
+f"""
+Description: This is the output of the positive case from the face recog 2023 algorithm.
+File: {frame_info.filename}
+Line: {frame_info.lineno}
+recogScore values:\n {recogScores}
+"""
+        )
+
+        return recogScores
+
+    def face_recog_2023_run_negative(self):
+        LABEL_FOR_PROGRESS_OF_TEST = "FaceRecog2023NegativeProg"
+
+        if self.pW:
+            self.pW.createProgressbar(LABEL_FOR_PROGRESS_OF_TEST)
+
+        userCount = len(self.users)
+        userFin = 0
+
+        testImageIndices = []
+        recogImageIndices = []
+        recogImageScores = []
+        usernames = []
+        recogUsernames = []
+
+        # reset previously used resources
+        self.FaceRecog2023NegativeUserTimer.clear()
+
+        for user_index,user_ in enumerate(self.users):
+            if self.pW:
+                self.pW.update(
+                    LABEL_FOR_PROGRESS_OF_TEST,
+                    "[FaceRecog2023Negative] Benchmarking User: {}/{}...".format(user_.username,self.userLimit),
+                    userFin/userCount * 100
+                )
+
+            if self.root:
+                self.root.update()
+
+            if len(user_.trainPictures) == 0 or len(user_.recogPictures) == 0:
+                print("WARNING: Training pictures empty!")
+                continue
+
+            self.FaceRecog2023NegativeUserTimer.startTimer(user_)
+
+            # negative case
+            ## Fetch User pictures
+            rand = random.randint(0,userCount - 1)
+            while rand == user_index:
+                rand = random.randint(0,userCount - 1)
+
+            imgs_train = self.users[rand].trainPictures.copy() # Images are flattened
+            imgs_test = user_.recogPictures.copy()
+
+            ## calculate scores
+            scores = self._execute_face_recog_2023_algo(imgs_train, imgs_test)
+
+            ## Prepare Score Dataframe
+            user_.recogScores = None # TODO: This is not used. Use it!
+            for testImageIndex in range(scores.shape[1]):
+                recognisedImageIndex = np.argmin(scores[:, testImageIndex])
+                recognisedImageScore = np.min(scores[:, testImageIndex])
+                testImageIndices.append(testImageIndex)
+                recogImageIndices.append(recognisedImageIndex)
+                recogImageScores.append(recognisedImageScore)
+                usernames.append(user_.username)
+                recogUsernames.append(self.users[rand].username)
+
+            userFin += 1
+            self.FaceRecog2023NegativeUserTimer.endTimer(user_)
+
+        if self.pW:
+            self.pW.finProgress(LABEL_FOR_PROGRESS_OF_TEST)
+
+        recogScores = pd.DataFrame(
+            data = {
+                'testImageIndex':  testImageIndices,
+                'username':        usernames, 
+                'recogImageIndex': recogImageIndices,
+                'recogUsername' :  recogUsernames,
+                'recogScore' :     recogImageScores
+            }).set_index('recogImageIndex').sort_values(by='testImageIndex')
+
+        frame_info = getframeinfo(currentframe())
+        print(
+f"""
+Description: This is the output of the negative case of the face recog 2023 algorithm.
 File: {frame_info.filename}
 Line: {frame_info.lineno}
 recogScore values:\n {recogScores}
